fix: fixed deployment for both contracts
