+++ stylus-dao/Cargo.toml
@@ -2,4 +2,5 @@
 [workspace]
 members = [
     "dao-contract",
+    "dao-token",
 ]
